Issue AP-5785: Workflow Coach: Recommending next best successor based on community usage statistics https://knime-com.atlassian.net/browse/AP-5785
Slightly nicer messages and link when no recommendations are displayed.
Now, also a link-text spanning more than one column allowed.
If 'reporting the usage data' is disabled, the link will say: "Node recommendations only available with usage data reporting."
A bit of restructuring of the inner code of the WorflowCoachView.