AP-5607: fixed inability to maximize editor pane immediately after launch
This is not the most satisfactory-feeling of fixes; the underlying
problem is that the CTabFolder instance for the editor pane does not get
its minimizable / maximizable set to true when the Eclipse browser
editor pane part gets added (if it never gets added, because we're
running from the SDK, for example, the addition of a WorkflowEditor
*does* result in the CTabFolder instance getting this set correctly.)

The reason a maximizing of a different view fixes this is because part
of the behavior of the e4 MinMaxAddon is to make sure the minimizable /
maximizable attributes are set to true on views that are being
collapsed.

This code change does not have an impact, i believe, on KNIME core
behaviour - but it will foil a third party developer who wants to have a
view that explicitly does not show the min/max controls for it (they
will be arguably foiled, though, the moment a user maximizes any other
view in their running application though.)

Overall, I would have preferred to have this tied into
KNIMEApplicationWorkbenchAdvisor, looking for a one-off set should it
see that the intro page has been created but in e4 i can currently find
no, even hacky, way to get access to the CTabFolder instances associated
to the IWorkbenchWindow / IWorkbenchPage.

Also changed:
	. variable shadowing in a constructor
	. fixed a classpath issue - o.k.w.help was pointing to Java6 libraries
despite having code that uses java.nio.file which was introduced in
Java7