AP-11680: 'Included-path-cache' updated even if metanode-templates in node repo is deactivated
Because we still want to be able to show the 'server-configured' and
'default' workflow groups (that contribute their templates to the node
repository) in the preference page. Gives the user more information to
decide whether to activate the function.

AP-11680 (Server config of workflow groups that contribute templates to
node repository)