AP-14496: For Linux/GTK, use SWT.TOOL for toolbar
Use SWT.TOOL (instead of SWT.ON_TOP) as a flag for the node annotation's
floating toolbar Shell. The other changes resolve bugs that became
visible
after this change: 1) The timeout workaround in ToolbarHideDecider can
be replaced by a proper solution and 2) Giving back focus to the text
input when a font size is selected will now happen *only* in that case.

AP-14496 (Extra click needed to edit a node's annotation)