AP-14568: Manually trigger show event for color dropdown
Manually trigger a SWT.Show event when a color dropdown is currently
open and another one is opened. This enables the floating toolbar to
properly recompute the clipping region, fixing AP-14568.

Because now show events are fired for the color dropdown in a way that
is in line with the intuition of TransientEditAssetManager, we have to
fix the bookeeping on currently shown elements, namely not add the
ColorDropDown multiple times to the same data structure.

AP-14568 (Workflow Annotation Edit Toolbar color chooser is buggy)