AP-15093: Prevent premature disposal of Color object
If a change to a workflow annotation was commited but the background
color had not changed, that Color object would still be set as the
current background color of the WorkflowAnnotationFigure but be disposed
beforehand under the handle `previousBackgroundColor`. On Linux, this
causes an error, as methods such as `setBackgroundColor` access the
current background color of the figure.

AP-15093 (On Linux, editing workflow annotation with a background colour throws error)