AP-15682: Use node extension manager to resolve nodes on dnd
... whenever possible. Not 100% possible for dynamic nodes atm because
the hub doesn't provide the node factory's additional settings. Thus,
we, first, fallback to use the node repository to resolve those nodes by
factory and node name. And, secondly, in case of deprecated dynamic
nodes we fallback to use the node extension manager, but not in an
efficient way.

AP-15682 (Deprecated nodes can't be drag'n'dropped from the hub)