AP-14591: Replace flag `APP_IS_UNDERGOING_EXIT`
The flag would not be cleared properly when the workflow save dialog is
canceled by the user, causing problems when intialising other workflows
later. This commit replaces the use of this flag with an
eclipse-internal way of determining whether the application is currently
closing.

AP-14591 (Opening any workflow in AP gives error)