AP-16655: Custom definition of mountpoint root for local file store
The local mountpoint root is usually the eclipse workspace directory.
However, if testflows are run with the TestflowRunnerApplication, those
are executed outside of the eclipse workspace. This causes, e.g.,
local, workflow-relative file/directory-accesses to fail (because the
workspace was assumed to be the root directory).
With this change, the TestflowRunnerApplication is able to set a custom
mountpoint root directory.

AP-16655 (TestflowRunner: fix workflow-relative reading/writing and enable component-download)