AP-17352: Layout editor browser can be provided via extension point
An extension point which is extended by a plugin in knime-cef in order
to be able to the CEF browser.
This introduced extension avoid the direct dependency from this
knime-workbench-plugin to knime-cef such that the CEF browser can still
be de-installed (which is at least required for 4.7.x; but not for
higher versions anymore).

AP-17352 (Provide CEF as an alternative to SWT for the layout editor)