AP-20541: Set new location in component projects after Save-As
This fixes a bug which led to the whole workspace being wiped.
The workflow editor (classic workbench) would assume the target location
is a temporary directory and clean it up because the workflow context
was not updated after a save-as operation.

AP-20541 (All local workflows get deleted by downloading a component from Hub)