DEVOPS-2351: register URIToFileResolve as declarative service (earlier registration)
After the eclipse bump we found that the service was added too late
during
testflow execution (in knime-buildworkflows). The service is/was only
registered on plugin activation and the plug-in was only accidently
activated
by some other extension point... which is too late.

Note the immediate="true" attribute.

DEVOPS-2351 (Update eclipse 2023-03 to 2024-03)